+++ src/django_integrity/conversion.py
@@ -4,7 +4,7 @@ import abc
 import contextlib
 import dataclasses
 import re
-from collections.abc import Iterator, Mapping
+from collections.abc import Iterator, Sequence
 
 from django import db as django_db
 
@@ -16,23 +16,26 @@ except ImportError:
 
 
 @contextlib.contextmanager
-def refine_integrity_error(rules: Mapping[_Rule, Exception]) -> Iterator[None]:
+def refine_integrity_error(
+    rules: Sequence[tuple[_Rule, Exception | type[Exception]]],
+) -> Iterator[None]:
     """
     Convert a generic IntegrityError into a more specific exception.
 
-    The conversion is based on a mapping of rules to exceptions.
+    The conversion is based on (rule, exception) pairs.
 
     Args:
-        rules: A mapping of rules to the exceptions we'll raise if they match.
+        rules: A sequence of rule, exception pairs.
+            If the rule matches the IntegrityError, the exception is raised.
 
     Raises:
-        An error from the rules mapping if an IntegrityError matches a rule.
+        The exception paired with the first matching rule.
         Otherwise, the original IntegrityError.
     """
     try:
         yield
     except django_db.IntegrityError as e:
-        for rule, refined_error in rules.items():
+        for rule, refined_error in rules:
             if rule.is_match(e):
                 raise refined_error from e
         raise